pyproject.toml
[tool.poetry]
name = "vector-search-mongodb"
version = "0.1.0"
description = ""
authors = ["Yu-Pang <tedchen524@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
pymongo = "^4.7.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
